+++ js/modules/map.js
@@ -2,9 +2,9 @@
 
 Map core functionality
 
-An object with options can be passed to the map:
+Customize Map Styles:
 
-
+https://snazzymaps.com/style/105672/django-cms-default
 
 Requires the API key variable in base_root.html:
 
@@ -39,53 +39,182 @@ export const initMap = function (options) {
     if (!options.map_styles) {
         options.map_styles = [
             {
-                "featureType":"water",
-                "elementType":"geometry.fill",
-                "stylers":[{"color":"#d3d3d3"}]
-            },{
-                "featureType":"transit",
-                "stylers":[{"color":"#808080"},{"visibility":"off"}]
-            },{
-                "featureType":"road.highway",
-                "elementType":"geometry.stroke",
-                "stylers":[{"visibility":"on"},{"color":"#b3b3b3"}]
-            },{
-                "featureType":"road.highway","elementType":"geometry.fill","stylers":[
-                {"color":"#ffffff"}
+                "featureType": "all",
+                "elementType": "labels.text.fill",
+                "stylers": [
+                    {
+                        "saturation": 36
+                    },
+                    {
+                        "color": "#333333"
+                    },
+                    {
+                        "lightness": 40
+                    }
                 ]
-            },{
-                "featureType":"road.local","elementType":"geometry.fill","stylers":[
-                {"visibility":"on"},{"color":"#ffffff"},{"weight":1.8}]
-            },{
-                "featureType":"road.local","elementType":"geometry.stroke","stylers":[{"color":"#d7d7d7"}]
-            },{
-                "featureType":"poi","elementType":"geometry.fill","stylers":[{"visibility":"on"},{"color":"#ebebeb"}]
-            },{
-                "featureType":"administrative","elementType":"geometry","stylers":[{"color":"#a7a7a7"}]
-            },{
-                "featureType":"road.arterial","elementType":"geometry.fill","stylers":[{"color":"#ffffff"}]
-            },{
-                "featureType":"road.arterial","elementType":"geometry.fill","stylers":[{"color":"#ffffff"}]
-            },{
-                "featureType":"landscape","elementType":"geometry.fill","stylers":[{"visibility":"on"},{"color":"#efefef"}]
-            },{
-                "featureType":"road","elementType":"labels.text.fill","stylers":[{"color":"#696969"}]
-            },{
-                "featureType":"administrative","elementType":"labels.text.fill","stylers":[{"visibility":"on"},{"color":"#737373"}]
-            },{
-                "featureType":"poi","elementType":"labels.icon","stylers":[{"visibility":"off"}]
-            },{
-                "featureType":"poi","elementType":"labels","stylers":[{"visibility":"off"}]
-            },{
-                "featureType":"road.arterial","elementType":"geometry.stroke","stylers":[{"color":"#d6d6d6"}]
-            },{
-                "featureType":"road","elementType":"labels.icon","stylers":[{"visibility":"off"}]
-            },{
-                "featureType":"poi","elementType":"geometry.fill","stylers":[{"color":"#dadada"}]
-            },{
-                "featureType":"landscape.natural","elementType":"labels","stylers":[{"visibility":"off"}]
+            },
+            {
+                "featureType": "all",
+                "elementType": "labels.text.stroke",
+                "stylers": [
+                    {
+                        "visibility": "on"
+                    },
+                    {
+                        "color": "#ffffff"
+                    },
+                    {
+                        "lightness": 16
+                    }
+                ]
+            },
+            {
+                "featureType": "all",
+                "elementType": "labels.icon",
+                "stylers": [
+                    {
+                        "visibility": "off"
+                    }
+                ]
+            },
+            {
+                "featureType": "administrative",
+                "elementType": "geometry.fill",
+                "stylers": [
+                    {
+                        "color": "#fefefe"
+                    },
+                    {
+                        "lightness": 20
+                    }
+                ]
+            },
+            {
+                "featureType": "administrative",
+                "elementType": "geometry.stroke",
+                "stylers": [
+                    {
+                        "color": "#fefefe"
+                    },
+                    {
+                        "lightness": 17
+                    },
+                    {
+                        "weight": 1.2
+                    }
+                ]
+            },
+            {
+                "featureType": "landscape",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#ebebeb"
+                    },
+                    {
+                        "lightness": 20
+                    }
+                ]
+            },
+            {
+                "featureType": "poi",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#f5f5f5"
+                    },
+                    {
+                        "lightness": 21
+                    }
+                ]
+            },
+            {
+                "featureType": "poi.park",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#dedede"
+                    },
+                    {
+                        "lightness": 21
+                    }
+                ]
+            },
+            {
+                "featureType": "road.highway",
+                "elementType": "geometry.fill",
+                "stylers": [
+                    {
+                        "color": "#ffffff"
+                    },
+                    {
+                        "lightness": 17
+                    }
+                ]
+            },
+            {
+                "featureType": "road.highway",
+                "elementType": "geometry.stroke",
+                "stylers": [
+                    {
+                        "color": "#ffffff"
+                    },
+                    {
+                        "lightness": 29
+                    },
+                    {
+                        "weight": 0.2
+                    }
+                ]
+            },
+            {
+                "featureType": "road.arterial",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#ffffff"
+                    },
+                    {
+                        "lightness": 18
+                    }
+                ]
+            },
+            {
+                "featureType": "road.local",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#ffffff"
+                    },
+                    {
+                        "lightness": 16
+                    }
+                ]
+            },
+            {
+                "featureType": "transit",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#f2f2f2"
+                    },
+                    {
+                        "lightness": 19
+                    }
+                ]
+            },
+            {
+                "featureType": "water",
+                "elementType": "geometry",
+                "stylers": [
+                    {
+                        "color": "#d1d8da"
+                    },
+                    {
+                        "lightness": 17
+                    }
+                ]
             }
-
         ];
     }